requirements.txt
django==5.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
social-auth-app-django==5.4.0
python-social-auth==0.3.6
python-jose==3.3.0
requests==2.31.0
Pillow